refactor: Redis 키 생성 및 주문번호 생성/사용 로직 리팩토링
